[Fix] 리뷰 요약 에러 수정

+++ src/main/java/com/fasttime/domain/review/dto/response/BootcampReviewSummaryDTO.java
@@ -3,5 +3,7 @@ package com.fasttime.domain.review.dto.response;
 public record BootcampReviewSummaryDTO(
     String bootcamp,
     double averageRating,
-    Long totalReviews
-) {}
+    int totalReviews
+) {
+
+}
